Fixes for EC2-Classic AutoScaling groups.
Improve the conversion of the autoscaling.LaunchConfiguration from the existing
AutoScaling group into a ec2.RequestSpotLaunchSpecification which is needed
for launching Spot instances. Some of the fields were incorrectly handled,
which was causing failures on EC2 Classic groups.

In addition, some of the previously skipped fields are now converted into their
closest equivalents.

Change-Id: I619672a001a6e642fea9a3aa975bc98158617703