Use CloudFormation for creating the Event Generator
Finally CloudFormation can create our Event Generator in the user's AWS
account, so the custom CloudFormation resource and its handler code are no
longer needed. They used to be needed for attaching to the SNS topic in
the author's AWS account, which was a temporary workaround for not being
able to create the Event Generator from CloudFormation.

This obsoletes a significant amount of code and testing data, and allows
reducing the IAM role permissions. It should also speed up the runtime
and reduce the bandwidth requirements, because the binary code may be
smaller, and we no longer need to prepare the API connections for SNS and
Lambda calls.