algo: Making the minimum on-demand value configurable
The configuration regarding the minimum instance number can be
configured from the CLI as well as from the tag of the ASG.

The priority goes as following:
* Tag - Strict number
* Tag - Percentage
* CLI - Strict number
* CLI - Percentage

Which means the CLI parameters can be overidden by the tags, but
percentages value can never overide the strict numbers - except if
the value is invalid (out of range, negative, etc). Which means,
that as soon as a good value is found the processing stops.

The percentage is the percentage of instances currently in the ASG,
not Max/Min/Desired. Please note that the values are rounded up,
so use it with logic.

If the number of On-Demand is above the minimum required, then the
processing of the ASG continues.
If the value is equal, then nothing needs to be done, the processing
stops.
If the value is below, then it means that a spot-instance need to be
terminated in order for the ELB to recreate an On-Demand instance.

I believe that with the current way of making the algorithm
configurable, it shouldn't be too complicated to implement other
elements.