tests: adding testing for new methods of the configurable algorithm
Almost all new methods have been covered up to one exception due to the
need of mocking ec2 calls (terminate). This hasn't been done, as a
proper testing 'environment'/package should be setup to be used in a
generic manner around the code.

Observation:
The different approaches of the AWS go-sdk seem to be whether using
interfaces, whether handlers, but in the current design using interfaces
might be more complex as it would require -dynamic- type assertions in
the code.