instance: fix some debug & type errors
Some error was done with typing pointer/nil/division by 0, etc.
The logic seems to be good, but once a chosenType is found, then all the
other take its place, despite being invalid, I wonder if this isn't a
bug due to the processing of ASG in the go routine.