instance: improve code coverage add basic timeout on tags
The tag method was basically an infinite loop if the calls were
returning an error, which is cleary not ideal. It has been slightly
modified in order to have a timeout - default to 5min -; this way we can
specify a shorter timeout for tests (5sec).

The implementation and logic isn't ideal, but allow to test that the
tagging method is working as expected.

Other methods from instanceManager have also been covered.