Pass LDFLAGS to the upstream version of Makefile.lambda (#136)
Previously the upstream Makefile.example didn't support LDFLAGS, which we needed in order to save the version string in the binaries.

Now that it does we can make use of it without maintaining a forked version. The custom version was a burden to maintain, it was breaking from time to time after upstream updates of the build Docker image.