add new global option on_demand_price_multiplier to support volume discounts
This patch adds a new global option called on_demand_price_multiplier
that is primarily intended to be used to support EC2 volume discounts.
The option could also be used to set your set your bid price to be
higher than the on demand price to reduce the chances that your spot
instances will be terminated.