Spot bids: Allow different policies to optimize costs
It is now possible to bid at a value based on the spot instance price
('aggressive' policy) rather than on-demand price ('normal' policy). Thus
allowing to bid for a price slightly higher than the spot price. This is
particularly useful to avoid having spot instances that cost too much.

To enable this, the policy has to be set as 'aggressive' ('normal' default) and
a 'buffer percentage' needs to be defined (10% higher default) to know how much
more users are willing to pay above the spot price found. Details about the
tags are provided in START.md.

Those parameters are configurable from Cloudfront/Terraform and the ASGs.