terraform: improve documentation for stack creation
When using latest terraform versions (>0.10) the terraform plugins
aren't integrated in the binary directly, so you need to install them
prior to running the creation.

The comment relative to the lambda binary has also been brought up to be
made more obvious that the .zip file is required to create the stack.