core: always use SecurityGroup IDs
We want to attach the correct SecurityGroups to newly created spot
instances. For that we just reuse the SecurityGroup attached to
the instances in the auto scaling group but it turns out that those
Security Groups can either be mentioned by name (in the classic EC2 or
the DefaultVPC case) or by id (in the Lunix/UNIX (VPC) case).

In order to make the handling of these different cases easier we try
to use IDs in all cases. If a SecurityGroup string starts with "sg-"
we assume it's an idea already and use it as is. Otherwise we search
for the string as a group name and use its ID.