Create notification event rule to trigger lambda across all regions
We cannot directly execute/trigger a lambda from a notification rule
due to same region restrictions imposed by AWS.
But a topic/subscription can call a lambda across regions. So we have
to create both these resources. EC2 -> EventRule -> Topic -> Lambda.
Another problem is we cannot get the topic ARN easily from AWS API as
they have not exposed to get a topic by its name. This can be solved
by creating the above resources(topic and event rule) inside a stack.
So a stack with an unique name is created and when autospotting stack
is deleted this dependent stack is also deleted across all regions.