Terminate unused spot instances #321
Currently.
A spotted instance is launched.
Before the next autospot execution the ondemand instance that autospot
next execution should have replaced is terminated.
The launched spot instance, that is no more necessary, is not terminated.

The idea is:
once we found a spot instance launched for an ASG, we first check if the
ASG still need it.
If not we terminate the spot instance.
Beacuse we cannot know if the ASG will need it in the future and after
how much time.