Termination Notification Configurable Action
Idea:
At the moment, when a termination notification message is received,
we are only supporting detach as action on the terminating spot instance.

Detaching an instance do not trigger Lifecycle Hook (LH).

This patch add a new option "termination_notification_action" with
relative CloudFormation parameter "TerminationNotificationAction" that
support the options:
detach
	current behaviour - no LH Triggered
terminate
	new behaviour - spot instance is terminated without decrementing
	desired capacity, so that ASG spawn a new instance - LH
	Triggered.
auto
	new behaviour - if ASG has a LH for Terminate Lifecycle
	Transition, action is "terminate", if not action is detach.

The option is global for all ASG.
But using "auto", behaviour is based on the presence of LH, this way it
can act in different ways.