Do not process AutoScaling groups while the CloudFormation stack that created them is in progress
Fixes for #13

When analyzing ASG to choose if process it or not
look for a tag named "aws:cloudformation:stack-name",
to know if the ASG is managed by CloudFormation.

If found, get the stack name from the value of the tag and find out
if CloudFormation is currently updating the stack.

In that case skip processing the ASG.