Rewrite main Handler method to parse sns and event
This way we can pass to main lambda a cloudwatch event or an sns
that contain/forward a cloudwatch's one.

The code is even more clear to read and implement a better logic.

As we now extract region directly from CW event, even in case of an sns
message, there is no more need to parse sns arn using strings library.

In a successive PR TerminationEventRuleFunction in regional_template.yaml
can be modified to direclty send CW event (no more need to encapsulate
it in an SNS message).