Take the Instance ID from the Launch Configuration
Running instances are sometimes not automatically replaced after
LaunchConfiguration changes, and the AMI used to launch the currently
running instances may be deleted even if instances launched using it are
still running.

We previously took the AMI ID from the instance we're trying to replace,
but if this instance is running from a deleted AMI, the spot instance will
fail to be launched. This is why it's better to just take the AMI ID
from the currently configured launch configuration.

In case of launch templates, the launch templates used on ASGs are expected
to mention the AMI, so we can just take it from there and not from the
running isntance we're trying to replace, where we may also run into the
same issue with deleted AMIs.