Implement event-based instance replacement
Context:

This change attempts to replace on-demand instances with spot as soon as they
are launched (triggered by the on-demand instance entering the pending state),
in order to avoid runnin expensive on-demand capacity in parallel with the
newly launched spot instances. This should cover new groups out of the box, as
well as scaling out or replacement of terminated spot instances.

It has a few nice side effects concerning cost, since on-demand instances are
likely to run for just a few seconds, as well as potentially tackling some
known issues with ElasticBeanstalk(#343, #344) and deletion of CloudFormation
stacks(#332).

It may also trick the ASG into running the start lifecycle hooks(#296), and
will immediately replace any scaled out instances, partially addressing #284.

The current/legacy replacement logic will still kick in when enabling
AutoSpotting for the first time on an existing group, as well as when spot
instances fail to be launched so the group must failover to on-demand capacity
in order to avoid downtime.

Once the legacy replacement logic is done on the current group, the group is
tagegd with a configuration tag that enables the event logic. This tag is also
set for all groups newer than an hour so their instances are automatically
replaced by the event-based replacement logic.

Current status
- changed regional CloudFormation stack to intercept EC2 instances
  as soon as entering the 'pending' and 'running' state'
- implement handling code for these events, that launches a replacement spot
  instance for each on-demand instance currently in pending state and performs
  the usual instance swapping logic when a spot instance enters the running
  state, without waiting for it to exit its grace period.
- largely kept the initial instance replacement logic as it is (waiting for
  the grace period) to avoid introducing regressions, and also in case the
  events are not properly propagated like in case of misconfigurations.
- it compiles and was executed against against real infrastructure and is
  working as expected for a couple of groups I tested it against.
- still missing test coverage for the new code and some existing tests were
  crashing and hat to be disabled for now
- implemented support for handling the AutoScaling lifecycle hooks when
  replacing new on-demand instances in event-based execution mode
- added support for running it locally against custom JSON event data in
  CloudFront Event format: https://docs.aws.amazon.com/AmazonCloudWatch/latest/events/EventTypes.html#ec2_event_type

ToDo:
- test against more infrastructure
- iron out any crashes
- write automated tests for the new code and re-enable older test cases so
  that coverage doesn't drop
- port regional CloudFormation stack changes to the main template so
  it can also run from a StackSet
- update the documentation to mention this new behavior