new param for belongsToEnabledASG to skip checking event enabled
shouldBeReplacedWithSpot call belongsToEnabledASG
shouldBeReplacedWithSpot is used both by handleNewOnDemandInstanceLaunch
 and swapWithGroupMember.

swapWithGroupMember is excuted when a spot launch message arrive on the
SQS Queue.
The message is sent only if ASG is enabled for event.
So there no need to check again it, even because with new workflow
 spot launch message are sent even if ASG is not enabled for event (Ex.
 cron or Lifecyclehook).
So i add a new boolean parameter, doNotCheckEventEnabled, to both shouldBeReplacedWithSpot and
belongsToEnabledASG to skip check for ASG enabled for event.

handleNewOnDemandInstanceLaunch was checking for both
shouldBeReplacedWithSpot and belongsToEnabledASG:
```
i.belongsToEnabledASG() && i.shouldBeReplacedWithSpot()
```

but as shouldBeReplacedWithSpot already have as condition
belongsToEnabledASG than line can be replaced with:
```
if i.shouldBeReplacedWithSpot(false) {
```